+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Thank you to all who have contributed!
 ### Added
 - Support aggregation function calls in the `ORDER BY` and `HAVING` clause
 - Add Support to transcribe `Overlaps` function
+- Add planner support to compare DATE and TIMESTAMP(Z)
 
 ### Changed